requirements.txt
Django==1.9.4
djangorestframework==3.3.3
django-filter=0.13.0
testfixtures==4.9.0
mock==1.0.1